+++ dropdbase_test/TableTests.cpp
@@ -591,7 +591,7 @@ TEST(TableTests, ClusteredIndexInsertWithNullValues_basic)
     table.CreateColumn("ColumnInt2", COLUMN_INT);
 
     std::unordered_map<std::string, std::any> data;
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask;
 
     std::vector<int32_t> dataInt1({2, 1, 5, 8, 102, 67, 5, 1});
     std::vector<int32_t> dataInt2({21, 12, 50, 80, 1020, 670, 60, 13});
@@ -599,8 +599,8 @@ TEST(TableTests, ClusteredIndexInsertWithNullValues_basic)
     data.insert({"ColumnInt1", dataInt1});
     data.insert({"ColumnInt2", dataInt2});
 
-    std::vector<int8_t> vectorMask1;
-    std::vector<int8_t> vectorMask2;
+    std::vector<int64_t> vectorMask1;
+    std::vector<int64_t> vectorMask2;
     vectorMask1.push_back(3);
     vectorMask2.push_back(11);
 
@@ -646,7 +646,7 @@ TEST(TableTests, ClusteredIndexInsertWithNullValues_advanced)
     table.CreateColumn("ColumnInt3", COLUMN_INT);
 
     std::unordered_map<std::string, std::any> data;
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask;
 
     std::vector<int32_t> dataInt1({9, 5, 5, 7, 12, 4, 8, 5});
     std::vector<int32_t> dataInt2({7, 5, 7, 5, 12, 89, 56, 7});
@@ -660,9 +660,9 @@ TEST(TableTests, ClusteredIndexInsertWithNullValues_advanced)
     data.insert({"ColumnInt2", dataInt2});
     data.insert({"ColumnInt3", dataInt3});
 
-    std::vector<int8_t> vectorMask1;
-    std::vector<int8_t> vectorMask2;
-    std::vector<int8_t> vectorMask3;
+    std::vector<int64_t> vectorMask1;
+    std::vector<int64_t> vectorMask2;
+    std::vector<int64_t> vectorMask3;
     vectorMask1.push_back(60);
     vectorMask2.push_back(90);
     vectorMask3.push_back(204);
@@ -1610,9 +1610,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_Int)
     data6.insert({"ColumnIntA", dataIntA6});
     data6.insert({"ColumnIntB", dataIntB6});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask;
-    std::vector<int8_t> vectorMaskA;
-    std::vector<int8_t> vectorMaskB;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask;
+    std::vector<int64_t> vectorMaskA;
+    std::vector<int64_t> vectorMaskB;
     vectorMaskA.push_back(3);
     vectorMaskB.push_back(0);
 
@@ -1651,9 +1651,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_Int)
     data7.insert({"ColumnIntA", dataIntA7});
     data7.insert({"ColumnIntB", dataIntB7});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask1;
-    std::vector<int8_t> vectorMaskA1;
-    std::vector<int8_t> vectorMaskB1;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask1;
+    std::vector<int64_t> vectorMaskA1;
+    std::vector<int64_t> vectorMaskB1;
     vectorMaskA1.push_back(0);
     vectorMaskB1.push_back(3);
 
@@ -2188,9 +2188,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_Long)
     data6.insert({"ColumnA", dataA6});
     data6.insert({"ColumnB", dataB6});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask;
-    std::vector<int8_t> vectorMaskA;
-    std::vector<int8_t> vectorMaskB;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask;
+    std::vector<int64_t> vectorMaskA;
+    std::vector<int64_t> vectorMaskB;
     vectorMaskA.push_back(3);
     vectorMaskB.push_back(0);
 
@@ -2229,9 +2229,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_Long)
     data7.insert({"ColumnA", dataA7});
     data7.insert({"ColumnB", dataB7});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask1;
-    std::vector<int8_t> vectorMaskA1;
-    std::vector<int8_t> vectorMaskB1;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask1;
+    std::vector<int64_t> vectorMaskA1;
+    std::vector<int64_t> vectorMaskB1;
     vectorMaskA1.push_back(0);
     vectorMaskB1.push_back(3);
 
@@ -2433,9 +2433,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_Float)
     data6.insert({"ColumnA", dataA6});
     data6.insert({"ColumnB", dataB6});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask;
-    std::vector<int8_t> vectorMaskA;
-    std::vector<int8_t> vectorMaskB;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask;
+    std::vector<int64_t> vectorMaskA;
+    std::vector<int64_t> vectorMaskB;
     vectorMaskA.push_back(3);
     vectorMaskB.push_back(0);
 
@@ -2472,9 +2472,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_Float)
     data7.insert({"ColumnA", dataA7});
     data7.insert({"ColumnB", dataB7});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask1;
-    std::vector<int8_t> vectorMaskA1;
-    std::vector<int8_t> vectorMaskB1;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask1;
+    std::vector<int64_t> vectorMaskA1;
+    std::vector<int64_t> vectorMaskB1;
     vectorMaskA1.push_back(0);
     vectorMaskB1.push_back(3);
 
@@ -2674,9 +2674,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_Double)
     data6.insert({"ColumnA", dataA6});
     data6.insert({"ColumnB", dataB6});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask;
-    std::vector<int8_t> vectorMaskA;
-    std::vector<int8_t> vectorMaskB;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask;
+    std::vector<int64_t> vectorMaskA;
+    std::vector<int64_t> vectorMaskB;
     vectorMaskA.push_back(3);
     vectorMaskB.push_back(0);
 
@@ -2713,9 +2713,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_Double)
     data7.insert({"ColumnA", dataA7});
     data7.insert({"ColumnB", dataB7});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask1;
-    std::vector<int8_t> vectorMaskA1;
-    std::vector<int8_t> vectorMaskB1;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask1;
+    std::vector<int64_t> vectorMaskA1;
+    std::vector<int64_t> vectorMaskB1;
     vectorMaskA1.push_back(0);
     vectorMaskB1.push_back(3);
 
@@ -2927,9 +2927,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_String)
     data6.insert({"ColumnA", dataA6});
     data6.insert({"ColumnB", dataB6});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask;
-    std::vector<int8_t> vectorMaskA;
-    std::vector<int8_t> vectorMaskB;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask;
+    std::vector<int64_t> vectorMaskA;
+    std::vector<int64_t> vectorMaskB;
     vectorMaskA.push_back(3);
     vectorMaskB.push_back(0);
 
@@ -2968,9 +2968,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_String)
     data7.insert({"ColumnA", dataA7});
     data7.insert({"ColumnB", dataB7});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask1;
-    std::vector<int8_t> vectorMaskA1;
-    std::vector<int8_t> vectorMaskB1;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask1;
+    std::vector<int64_t> vectorMaskA1;
+    std::vector<int64_t> vectorMaskB1;
     vectorMaskA1.push_back(0);
     vectorMaskB1.push_back(3);
 
@@ -3210,9 +3210,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_Point)
     data6.insert({"ColumnA", dataA6});
     data6.insert({"ColumnB", dataB6});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask;
-    std::vector<int8_t> vectorMaskA;
-    std::vector<int8_t> vectorMaskB;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask;
+    std::vector<int64_t> vectorMaskA;
+    std::vector<int64_t> vectorMaskB;
     vectorMaskA.push_back(3);
     vectorMaskB.push_back(0);
 
@@ -3255,9 +3255,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_Point)
     data7.insert({"ColumnA", dataA7});
     data7.insert({"ColumnB", dataB7});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask1;
-    std::vector<int8_t> vectorMaskA1;
-    std::vector<int8_t> vectorMaskB1;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask1;
+    std::vector<int64_t> vectorMaskA1;
+    std::vector<int64_t> vectorMaskB1;
     vectorMaskA1.push_back(0);
     vectorMaskB1.push_back(3);
 
@@ -3535,9 +3535,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_Polygon)
     data6.insert({"ColumnA", dataA6});
     data6.insert({"ColumnB", dataB6});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask;
-    std::vector<int8_t> vectorMaskA;
-    std::vector<int8_t> vectorMaskB;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask;
+    std::vector<int64_t> vectorMaskA;
+    std::vector<int64_t> vectorMaskB;
     vectorMaskA.push_back(3);
     vectorMaskB.push_back(0);
 
@@ -3582,9 +3582,9 @@ TEST(TableTests, InsertInto_IsUnique_CreateColumnWithConstraint_Polygon)
     data7.insert({"ColumnA", dataA7});
     data7.insert({"ColumnB", dataB7});
 
-    std::unordered_map<std::string, std::vector<int8_t>> nullMask1;
-    std::vector<int8_t> vectorMaskA1;
-    std::vector<int8_t> vectorMaskB1;
+    std::unordered_map<std::string, std::vector<int64_t>> nullMask1;
+    std::vector<int64_t> vectorMaskA1;
+    std::vector<int64_t> vectorMaskB1;
     vectorMaskA1.push_back(0);
     vectorMaskB1.push_back(3);
